# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Quorum — Multi-tenant Board of Advisors API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /mobile-auth/token-exchange:
    post:
      operationId: exchangeMobileAuthorizationCode
      tags: [auth]
      summary: Mobile token exchange
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ExchangeMobileAuthorizationCodeBody" }
      responses:
        "200":
          description: Session token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ExchangeMobileAuthorizationCodeResponse" }
...
  /tenants/{tenantId}/boards:
    get:
      operationId: listBoards
      tags: [boards]
      parameters:
        - $ref: "#/components/parameters/TenantIdParam"
      responses:
        "200":
          description: Boards
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/BoardSummary" }
    post:
      operationId: createBoard
      tags: [boards]
      parameters:
        - $ref: "#/components/parameters/TenantIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateBoardBody" }
      responses:
        "201":
          description: Board created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Board" }
  /boards/{boardId}:
    get:
      operationId: getBoard
      tags: [boards]
      parameters:
        - $ref: "#/components/parameters/BoardIdParam"
      responses:
        "200":
          description: Board with members
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BoardDetail" }
    patch:
      operationId: updateBoard
      tags: [boards]
      parameters:
        - $ref: "#/components/parameters/BoardIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateBoardBody" }
      responses:
        "200":
          description: Updated board
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Board" }
    delete:
      operationId: deleteBoard
      tags: [boards]
      parameters:
        - $ref: "#/components/parameters/BoardIdParam"
      responses:
        "204":
          description: Deleted
...
  /grounding-documents:
    post:
      operationId: registerGroundingDocument
      tags: [grounding]
      summary: Register an uploaded grounding document and extract text
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterGroundingDocumentBody" }
      responses:
        "201":
          description: Document registered
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GroundingDocument" }
  /grounding-documents/{documentId}:
    get:
      operationId: getGroundingDocument
      tags: [grounding]
      parameters:
        - $ref: "#/components/parameters/DocumentIdParam"
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GroundingDocument" }
...
  /boards/{boardId}/sessions:
    get:
      operationId: listSessions
      tags: [sessions]
      parameters:
        - $ref: "#/components/parameters/BoardIdParam"
      responses:
        "200":
          description: Sessions
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SessionSummary" }
    post:
      operationId: createSession
      tags: [sessions]
      parameters:
        - $ref: "#/components/parameters/BoardIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateSessionBody" }
      responses:
        "201":
          description: Session created (status=running)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionSummary" }
...
components:
  parameters:
    TenantIdParam:
      name: tenantId
      in: path
      required: true
      schema: { type: string }
    BoardIdParam:
      name: boardId
      in: path
      required: true
      schema: { type: string }
...
    DocumentIdParam:
      name: documentId
      in: path
      required: true
      schema: { type: string }
...
  schemas:
...
    ExchangeMobileAuthorizationCodeBody:
      type: object
      properties:
        code: { type: string, minLength: 1 }
        code_verifier: { type: string, minLength: 1 }
        redirect_uri: { type: string, minLength: 1, format: uri }
        state: { type: string, minLength: 1 }
        nonce: { type: ["string", "null"] }
      required: [code, code_verifier, redirect_uri, state]
    ExchangeMobileAuthorizationCodeResponse:
      type: object
      properties:
        token: { type: string }
      required: [token]
...
    BoardSize:
      type: integer
      enum: [3, 5, 7, 9]
    SessionMode:
      type: string
      enum: [ADVISORY, BOARD, REVIEW]
    SessionStatus:
      type: string
      enum: [running, complete, failed]
...
    BoardSummary:
      type: object
      properties:
        id: { type: string }
        tenantId: { type: string }
        name: { type: string }
        description: { type: ["string", "null"] }
        topicArea: { type: ["string", "null"] }
        size: { $ref: "#/components/schemas/BoardSize" }
        memberCount: { type: integer }
        sessionCount: { type: integer }
        lastSessionAt: { type: ["string", "null"], format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, tenantId, name, size, memberCount, sessionCount, updatedAt]
    Board:
      type: object
      properties:
        id: { type: string }
        tenantId: { type: string }
        name: { type: string }
        description: { type: ["string", "null"] }
        topicArea: { type: ["string", "null"] }
        masterInstructionsText: { type: string }
        size: { $ref: "#/components/schemas/BoardSize" }
        defaultMemberModel: { type: string }
        defaultMasterModel: { type: string }
        temperature: { type: number }
        createdBy: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, tenantId, name, masterInstructionsText, size, defaultMemberModel, defaultMasterModel, temperature, createdAt, updatedAt]
    BoardMember:
      type: object
      properties:
        id: { type: string }
        boardId: { type: string }
        name: { type: string }
        roleTitle: { type: string }
        lensDescription: { type: ["string", "null"] }
        instructionsText: { type: string }
        groundingDocumentId: { type: ["string", "null"] }
        groundingDocument:
          oneOf:
            - { $ref: "#/components/schemas/GroundingDocument" }
            - { type: "null" }
        modelOverride: { type: ["string", "null"] }
        ordering: { type: integer }
        createdAt: { type: string, format: date-time }
      required: [id, boardId, name, roleTitle, instructionsText, ordering, createdAt]
    BoardDetail:
      allOf:
        - $ref: "#/components/schemas/Board"
        - type: object
          properties:
            members:
              type: array
              items: { $ref: "#/components/schemas/BoardMember" }
          required: [members]
    CreateBoardBody:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        description: { type: ["string", "null"] }
        topicArea: { type: ["string", "null"] }
        size: { $ref: "#/components/schemas/BoardSize" }
        masterInstructionsText: { type: ["string", "null"] }
      required: [name, size]
    UpdateBoardBody:
      type: object
      properties:
        name: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        topicArea: { type: ["string", "null"] }
        masterInstructionsText: { type: ["string", "null"] }
        size:
          oneOf:
            - { $ref: "#/components/schemas/BoardSize" }
            - { type: "null" }
        defaultMemberModel: { type: ["string", "null"] }
        defaultMasterModel: { type: ["string", "null"] }
        temperature: { type: ["number", "null"] }
...
    GroundingDocument:
      type: object
      properties:
        id: { type: string }
        tenantId: { type: string }
        filename: { type: string }
        contentType: { type: string }
        storagePath: { type: string }
        characterCount: { type: integer }
        truncated: { type: boolean }
        uploadedAt: { type: string, format: date-time }
      required: [id, tenantId, filename, contentType, storagePath, characterCount, truncated, uploadedAt]
    RegisterGroundingDocumentBody:
      type: object
      properties:
        tenantId: { type: string }
        objectPath: { type: string }
        filename: { type: string }
        contentType: { type: string }
        size: { type: integer }
      required: [tenantId, objectPath, filename, contentType, size]
...
    SessionSummary:
      type: object
      properties:
        id: { type: string }
        boardId: { type: string }
        mode: { $ref: "#/components/schemas/SessionMode" }
        questionText: { type: string }
        status: { $ref: "#/components/schemas/SessionStatus" }
        startedAt: { type: string, format: date-time }
        completedAt: { type: ["string", "null"], format: date-time }
        totalCostCents: { type: ["integer", "null"] }
      required: [id, boardId, mode, questionText, status, startedAt]
...
    CreateSessionBody:
      type: object
      properties:
        mode: { $ref: "#/components/schemas/SessionMode" }
        questionText: { type: string, minLength: 1, maxLength: 8000 }
        allHands: { type: boolean }
      required: [mode, questionText]
...