# lib/api-zod/src/generated/types/registerGroundingDocumentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quorum — Multi-tenant Board of Advisors API
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterGroundingDocumentBody {
  tenantId: string;
  objectPath: string;
  filename: string;
  contentType: string;
  size: number;
}